Move tracer to a new dir
